pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "tracemaid"
version = "0.1.3"
description = "OpenTelemetry trace analysis and Mermaid diagram generation"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "KR", email = "Karthickrajam18@gmail.com"}
]
keywords = ["opentelemetry", "tracing", "mermaid", "visualization", "observability"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: System :: Monitoring",
]
requires-python = ">=3.9"
dependencies = [
    "numpy>=1.21.0",
    "scipy>=1.7.0",
    "scikit-learn>=1.0.0",
    "opentelemetry-api>=1.20.0",
    "opentelemetry-sdk>=1.20.0",
]

[project.optional-dependencies]
fastapi = [
    "opentelemetry-instrumentation-fastapi>=0.41b0",
    "opentelemetry-instrumentation-logging>=0.41b0",
]
instrumentation = [
    "opentelemetry-instrumentation-requests>=0.41b0",
    "opentelemetry-instrumentation-httpx>=0.41b0",
    "opentelemetry-instrumentation-sqlalchemy>=0.41b0",
    "opentelemetry-instrumentation-redis>=0.41b0",
]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "fastapi>=0.100.0",
    "uvicorn>=0.23.0",
    "httpx>=0.24.0",
]
all = [
    "tracemaid[fastapi,instrumentation,dev]",
]

[project.scripts]
tracemaid = "tracemaid.cli:main"

[project.urls]
Homepage = "https://github.com/karthyick/tracemaid"
Repository = "https://github.com/karthyick/tracemaid"
Issues = "https://github.com/karthyick/tracemaid/issues"

[tool.setuptools.packages.find]
where = ["."]
include = ["tracemaid*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = "-v --tb=short"

[tool.coverage.run]
source = ["tracemaid"]
branch = true

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
]
